MT D*Lite: makes the algorithm work when the robot moves, but fails whenever it moves too much.
Need to not choose goals as obstacles, and need to fix optimized deletion deleting seemingly everything including the subtree rooted at s_start

Restructured code so it's nice and in its own folder
