+++ igvc_ws/src/igvc_msgs/CMakeLists.txt
@@ -62,11 +62,10 @@ add_message_files(
 )
 
 ## Generate services in the 'srv' folder
-# add_service_files(
-#   FILES
-#   Service1.srv
-#   Service2.srv
-# )
+add_service_files(
+  FILES
+  EKFService.srv
+)
 
 ## Generate actions in the 'action' folder
 # add_action_files(
