+++ igvc_ws/src/igvc_nav/src/mt_dstar_node.py
@@ -4,9 +4,10 @@ import rospy
 from std_msgs.msg import String
 from nav_msgs.msg import OccupancyGrid
 from igvc_msgs.msg import motors, ekf_state
+from igvc_msgs.srv import EKFService
 import copy
 import numpy as np
-from mt_dstar_lite import mt_dstar_lite
+from path_planner.mt_dstar_lite import mt_dstar_lite
 
 motor_pub = rospy.Publisher("/igvc/motors_raw", motors, queue_size=10)
 path_pub = rospy.Publisher("/igvc_nav/path_map", OccupancyGrid, queue_size=10)
@@ -16,32 +17,12 @@ map_init = False
 planner = mt_dstar_lite()
 
 # Localization tracking
-ekf_offset = (0, 0)
 prev_coord = (0, 0)
-ekf_init = False
-GRID_SIZE = 0.1     # Map block size in meters
-
-
-# TODO: Do this as a service instead so we can get the EKF offset when the map updates rather than accumulating it ourselves
-def ekf_callback(ekf_data):
-    global prev_coord, ekf_init, ekf_offset
-
-    # Extract coordinates
-    xk = ekf_data.x_k
-    x = xk[3]
-    y = xk[4]
-
-    # Determine offset (in blocks)
-    ekf_offset = ((x - prev_coord[0]) / GRID_SIZE, (y - prev_coord[1]) / GRID_SIZE)
-
-    # Set the EKF to be initialized
-    ekf_init = True
-# TODO: Do this as a service instead so we can get the EKF offset when the map updates rather than accumulating it ourselves
-
-
+GRID_SIZE = 0.01     # Map block size in meters
+#TODO: the grid size should be 0.1 I think
 
 def c_space_callback(c_space):
-    global planner, map_init, ekf_init
+    global planner, map_init
 
     # Get the grid data
     grid_data = c_space.data
@@ -81,14 +62,19 @@ def c_space_callback(c_space):
 
     # MOVING TARGET D*LITE
     # If this is the first time receiving a map, initialize the path planner and plan the first path
-    if True: #map_init is False:
+    if map_init == False:
         planner.initialize(200, 200, (100, 100), best_pos, cost_map)
         path = planner.plan()
+        map_init = True
     # Otherwise, replan the path
     else:
-        # TODO: figure out how to do this
+        # Get the EKF's robot state estimate
+        robot_state = rospy.ServiceProxy('/igvc_ekf/get_robot_state', EKFService)
+        xk = robot_state()
+        robot_pos = (100 - int(xk.state.x_k[3] / GRID_SIZE), 100 + int(xk.state.x_k[4] / GRID_SIZE))
+        print(robot_pos)
         # Request the planner replan the path
-        path = planner.replan(config_space.tolist(), (100, 100), best_pos, cost_map)
+        path = planner.replan(robot_pos, best_pos, cost_map)
 
 
     print "path:"
@@ -120,7 +106,9 @@ def mt_dstar_node():
 
     # Subscribe to necessary topics
     rospy.Subscriber("/igvc_slam/config_space", OccupancyGrid, c_space_callback, queue_size=1)  # Mapping
-    rospy.Subscriber("/igvc_ekf/filter_output", ekf_state, ekf_callback, queue_size = 1)        # Localization
+
+    # Wait for the EKF to start advertising its service
+    rospy.wait_for_service('/igvc_ekf/get_robot_state')
 
     # Wait for topic updates
     rospy.spin()
